ai-tuner/requirements.txt
numpy==1.24.3
pandas==2.0.2
flask==2.3.2
prometheus-client==0.17.1
prometheus-api-client==0.5.4
pyyaml==6.0.1
requests==2.31.0